[REF] Refactored MyInfo project
