+++ SecondBlock/my-info/src/components/MyInfo.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+function MyInfo() {
+  return (
+    <div className="widget-sidebar">
+      <h1 className="title-s"> Raccoon Developer </h1>
+      <p>Blurb about myself. Coding...</p>
+      <ul className="list-ico">
+        <span>Alps</span>
+        <span>Chzeh</span>
+        <span>Tokio</span>
+      </ul>
+    </div>
+  )
+}
+
+export default MyInfo
