+++ SecondBlock/my-info/src/index.js
@@ -10,20 +10,9 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
+import MyInfo from './components/MyInfo'
 import './style-green.css';
 
-function MyInfo() {
-  return (
-    <div className="widget-sidebar">
-      <h1 className="title-s"> Raccoon Developer </h1>
-      <p>Blurb about myself. Coding...</p>
-      <ul className="list-ico">
-        <span>Alps</span>
-        <span>Chzeh</span>
-        <span>Tokio</span>
-      </ul>
-    </div>
-  )
-}
+
 
 ReactDOM.render(<MyInfo />, document.getElementById('root'));
